requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cachetools==4.0.0
certifi==2018.11.29
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
Django==2.1.5
django-bootstrap3==11.0.0
django-extensions==3.0.9
django-filter==2.2.0
djangorestframework==3.11.0
enum34==1.1.6
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.7
Jinja2==2.10
Markdown==3.2
MarkupSafe==1.1.0
multidict==4.5.2
oauth2client==4.1.3
oauthlib==3.1.0
peewee==3.13.1
Pillow==7.2.0
psycopg2==2.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pygsheets==2.0.3
python-telegram-bot==12.4.2
pytz==2018.9
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.0
six==1.11.0
tornado==6.0.3
uritemplate==3.0.1
urllib3==1.24.1
virtualenv==16.7.9
vk-api==11.3.0
yarl==1.3.0
